Moving Decoding to Common
