+++ Unosquare.FFME.Common/Decoding/AudioComponent.cs
