+++ Unosquare.FFME.Common/Decoding/ClosedCaptions/CCColorType.cs
