+++ Unosquare.FFME.Common/Decoding/ClosedCaptions/CCStyleType.cs
