+++ Unosquare.FFME.Common/Decoding/ClosedCaptions/ClosedCaptionCollection.cs
