+++ Unosquare.FFME.Common/Decoding/HardwareAccelerator.cs
