+++ Unosquare.FFME.Common/Decoding/MediaComponent.cs
