+++ Unosquare.FFME.Common/Decoding/SeekRequirement.cs
