+++ Unosquare.FFME.Common/Decoding/SubtitleBlock.cs
