+++ Unosquare.FFME.Common/Decoding/SubtitleComponent.cs
