+++ Unosquare.FFME.Common/Decoding/SubtitleFrame.cs
