+++ Unosquare.FFME.Common/Decoding/VideoComponent.cs
