+++ Unosquare.FFME.Common/Decoding/VideoFrame.cs
