+++ Unosquare.FFME.Common/MediaContainerException.cs
