+++ Unosquare.FFME.Common/MediaInfo.cs
