+++ Unosquare.FFME.Common/MediaOptions.cs
